Separated model error estimation from planning
- some tweaks
